Steering behaviour demonstration
Using a steering and a fleeing force, this example manipulates a path of points which are represented by a string of text. As the text is hovered the flee behaviour is activated alongside the steering behaviour which tries to move the point back into position so long as the fleeing behaviour isn't stronger than the steering one.